mcp-server/requirements.txt
mcp
fastapi
uvicorn
PyJWT
httpx
pydantic
sse-starlette
